refactor(assign-proxies): factor out ProxyGraph and document in readme
